project_feedback_ckasahara
This topic is very interesting! You did a great job explaining why this topic is necessary to look at. Since you are comparing income to the proximity to supermarket, I think it would be interesting to differentiate the type of supermarket across Buffalo (ex. Wegmans, Tops, Walmart, etc.). It isn't necessary, but it would add another interesting layer to the topic that you are addressing. Overall, very solid project!